+++ engine/model/src/main/scala/server/life/MonsterDropEntry.scala
@@ -1,7 +1,7 @@
 package server.life
 
 class MonsterDropEntry(private var _itemId: Int, private var _chance: Int, private var _minimum: Int,
-                       private var _maximum: Int, private var _questId: Short) {
+                       private var _maximum: Int, private var _questId: Int) {
   def itemId: Int = _itemId
 
   def chance: Int = _chance
@@ -10,5 +10,5 @@ class MonsterDropEntry(private var _itemId: Int, private var _chance: Int, priva
 
   def maximum: Int = _maximum
 
-  def questId: Short = _questId
+  def questId: Int = _questId
 }
