+++ engine/model/src/main/scala/server/life/MonsterGlobalDropEntry.scala
@@ -1,7 +1,7 @@
 package server.life
 
 class MonsterGlobalDropEntry(private var _itemId: Int, private var _chance: Int, private var _continentId: Int,
-                             private var _minimum: Int, private var _maximum: Int, private var _questId: Short) {
+                             private var _minimum: Int, private var _maximum: Int, private var _questId: Int) {
   def itemId: Int = _itemId
 
   def chance: Int = _chance
@@ -12,5 +12,5 @@ class MonsterGlobalDropEntry(private var _itemId: Int, private var _chance: Int,
 
   def maximum: Int = _maximum
 
-  def questId: Short = _questId
+  def questId: Int = _questId
 }
